api/services/market/requirements.txt
uvicorn
fastapi
pydantic
psycopg2-binary